Multi Server Executor
Currently this component only supports a single DNS server so with #6
in mind I started working on adding multiple server support. But ran
into the issue that if the first server on the list isn't responding it
takes a while before falling over the the next server in the list. So
with my work on https://github.com/reactphp/socket/pull/196 in mind and
specifically the basic principle of
https://tools.ietf.org/html/rfc8305#section-3.1 the work shifted to
sending the queries out with a delay between them rather then waiting
for the first server to fail/succeed. On the plus side records will
resolve faster when a server in the list is failing. On the down side it
 will send more queries thus generating a bit more network traffic.