Add createFromConfig factory methods
This little bit of syntactical sugar makes it cleaner to use the Config
class to load the system resolver configuration. Which right now is
limited to only one nameserver. By introducing this factory method
completing the implementation in #131 will also allow for instant
multiple server support out of the box without any need for users to
change code on their end to implement it.